requirements.txt
# Scraping
scrapy>=2.11.0

# Data Processing
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-pinecone>=0.0.1
langchain-community>=0.1.0

# Vector Database
pinecone-client>=3.0.0

# Embeddings & LLM
openai>=1.12.0
sentence-transformers>=2.5.0

# Utilities
python-dotenv>=1.0.0

# Pandas
pandas>=1.5.3

# Testing
pytest>=7.0.0
